Feature/better notification (#5)
* feat: add foundational files, workflows. improve alert functionality

- Introduced `AGENTS.md` for AI agent instructions and coding guidelines.
- Added `CONTRIBUTING.md` to outline contribution processes and expectations.
- Created `SECURITY.md` to define the security policy and reporting procedures.
- Added example configuration file `config.example.json` for user reference.
- Established issue templates for bug reports, feature requests, and questions in the `.github/ISSUE_TEMPLATE` directory.
- Implemented a release guide in `RELEASE_GUIDE.md` to assist with versioning and release processes.
- Updated `README.md` with troubleshooting and monitoring sections.
- Added CI workflow configuration in `.github/workflows/ci.yml` for automated testing and checks.
- Included various notifier implementations and tests to enhance notification capabilities.

* refactor: enhance logging and build process in Makefile

- Updated Makefile to include a 'clean' target and specify output binary name for the build.
- Changed log level from WARN to INFO for startup and heartbeat messages in main.go.
- Improved logging in monitor.go to provide more detailed information about aircraft detection and notifications.
- Added new logging levels (DEBUG, INFO) and corresponding functions in logger.go.
- Expanded logger tests to cover various log levels and scenarios.

* chore: update dependencies and improve security checks

- Updated Go version in CI workflows from 1.23 to 1.24 for better compatibility.
- Modified Makefile to install gosec if not already present, enhancing security scanning setup.
- Added go.uber.org/goleak dependency to go.sum for improved testing capabilities.
- Enhanced logger tests to ensure proper closure of write pipes, preventing potential resource leaks.
- Added ReadHeaderTimeout to HTTP server configuration in test server to mitigate Slowloris attacks.

* chore: update gosec installation and execution in Makefile and CI workflow

- Modified Makefile to install gosec in the current directory and execute it from there.
- Updated CI workflow to reflect the new installation path for gosec, ensuring consistent execution during security scans.
- Cleaned up the Makefile by adding a target to remove the gosec binary during the clean process.

* chore: replace gosec with govulncheck in Makefile and CI workflow

- Updated Makefile to install and run govulncheck for enhanced security scanning.
- Modified CI workflow to install govulncheck and execute it, removing references to gosec.
- Cleaned up the Makefile by removing the gosec binary from the clean target.

+++ internal/piaware/piaware.go
@@ -20,7 +20,7 @@ type Aircraft struct {
 
 // Data represents the piaware aircraft JSON response.
 type Data struct {
-	Now      int64      `json:"now"`
+	Now      float64    `json:"now"`
 	Aircraft []Aircraft `json:"aircraft"`
 }
